Separate RGBFIX header fixing from CLI option parsing
