+++ include/fix/fix.hpp
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: MIT
+
+#ifndef RGBDS_FIX_FIX_HPP
+#define RGBDS_FIX_FIX_HPP
+
+bool fix_ProcessFile(char const *name, char const *outputName);
+
+#endif // RGBDS_FIX_FIX_HPP
